requirements.txt
scikits-bootstrap==1.0.1
pandas==1.2.5
scikit-learn==0.22.1
pytorch==1.8.0
matplotlib==3.1.3
pytest==5.3.5
